add the two properties endpoints
